refactor(backend): separate routes into REST and Websocket Routes #30
